+++ src/main/java/org/example/fibonacci/FibonacciIterative.java
@@ -6,7 +6,7 @@ import static org.example.number.Constants.MATH_CONTEXT;
 
 public class FibonacciIterative implements FibonacciStrategy<BigDecimal> {
     @Override
-    public BigDecimal solveFibonacci(int n) {
+    public BigDecimal solveFibonacci(long n) {
         if (n <= 1) {
             return BigDecimal.valueOf(n);
         }
